mappersは以下のクラスを整理
